test: add initial unittest implementation
including a unittest for filterOut() in actions.go and an integration
test in writing bytes from main.go's run() into a *bytes.Buffer{} to
check output

added some generic testcases that checks if expected output is received
with the current available flags. with combinations

aded some test files to 'walk' on in an idiomatic testdata directory.